url2pkg: clean and extract after adjusting the package
The first 'bmake extract' is done only to look at the files from the
package, to determine the dependencies.  In that stage, the package
cannot be built successfully.  To make the whole workflow less
surprising, install the auto-detected dependencies before leaving the
rest of the work to the human package maintainer.  This means that
running url2pkg produces more output and may take significantly more
time, but that time would be spent anyway later, so it's a net win.

Suggested by gdt@.