mariadb*-client: Force HAVE_CURSES_H when using ncurses.
For some reason mariadb removed any cmake detection for HAVE_CURSES_H so it is
never defined, but the source is still littered with it, causing compilation
problems because curses.h never gets included.

The whole thing is incredibly messy, and this is the least likely to
accidentally break some alternate configuration, but implementations that use
curses.h which isn't ncurses are likely to still be broken.